backend/requirements.txt
flask
streamlit
pypdf2
sentence-transformers
faiss-cpu
python-docx
python-dotenv
requests
openai
